+++ tests/testthat/test_stats.R
@@ -63,10 +63,9 @@ context("Logistic regression (logistic)")
 test_that("logistic", {
   result <- logistic(titanic, "survived", c("pclass", "sex"))
   res1 <- capture.output(summary(result))[13] %>% trim()
-  # cat(paste0(res1, "\n"))
-  res2 <- "sex|male    0.080      -2.522     0.163 -15.447  < .001 ***"
+  cat(paste0(res1, "\n"))
+  res2 <- "sex|male    0.080 -92.0%      -2.522     0.163 -15.447  < .001 ***"
   expect_equal(res1, res2)
-
   result <- logistic(titanic, "survived", "pclass:sex")
   res1 <- capture.output(summary(result))[13] %>% trim()
   expect_equal(res1, res2)